move signal models to own file
